move all2all to utils
